+++ process/process_others.go
@@ -1,3 +1,4 @@
+//go:build !windows
 // +build !windows
 
 // Copyright © 2017-2023 Wei Shen <shenwei356@gmail.com>
@@ -27,9 +28,7 @@ import (
 	"bytes"
 	"context"
 	"errors"
-	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"strconv"
@@ -41,17 +40,13 @@ import (
 
 func getProcess(pid int) (processHandle int, processExists bool, accessGranted bool, err error) {
 	processHandle = INVALID_HANDLE
-	processExists = doesProcessExist(pid)
-	accessGranted = false
-	if processExists {
-		// check if we have access to send a signal to the process
-		pidStr := strconv.Itoa(pid)
-		_, err = exec.Command("kill", "-0", pidStr).Output()
-		if err == nil {
-			// just use the pid as the process handle
-			processHandle = pid
-			accessGranted = true
-		}
+	err = syscall.Kill(pid, 0)
+	if err == nil {
+		processHandle = pid
+		processExists = true
+		accessGranted = true
+	} else if err == syscall.EPERM {
+		processExists = true
 	}
 	return processHandle, processExists, accessGranted, err
 }
@@ -61,16 +56,8 @@ func releaseProcessByHandle(processHandle int) {
 }
 
 func doesProcessExist(processHandle int) (processExists bool) {
-	processExists = false
-	// check if the process exists
-	// just use the process handle as the pid
-	pid := processHandle
-	pidStr := strconv.Itoa(pid)
-	_, err := exec.Command("ps", "-p", pidStr).Output()
-	if err == nil {
-		processExists = true
-	}
-	return processExists
+	err := syscall.Kill(processHandle, 0)
+	return err == nil || err == syscall.EPERM
 }
 
 // from https://github.com/cloudfoundry/gosigar/blob/master/sigar_linux.go
@@ -82,7 +69,7 @@ var system struct {
 var Procd string
 
 func readFile(file string, handler func(string) bool) error {
-	contents, err := ioutil.ReadFile(file)
+	contents, err := os.ReadFile(file)
 	if err != nil {
 		return err
 	}
@@ -127,7 +114,7 @@ func procFileName(pid int, name string) string {
 
 func readProcFile(pid int, name string) ([]byte, error) {
 	path := procFileName(pid, name)
-	contents, err := ioutil.ReadFile(path)
+	contents, err := os.ReadFile(path)
 
 	if err != nil {
 		if perr, ok := err.(*os.PathError); ok {
@@ -166,11 +153,26 @@ func getShell() (shell string) {
 }
 
 func getCommand(ctx context.Context, qcmd string) *exec.Cmd {
+	var command *exec.Cmd
 	if ctx != nil {
-		return exec.CommandContext(ctx, getShell(), "-c", qcmd)
+		command = exec.CommandContext(ctx, getShell(), "-c", qcmd)
 	} else {
-		return exec.Command(getShell(), "-c", qcmd)
+		command = exec.Command(getShell(), "-c", qcmd)
 	}
+	command.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
+	if ctx != nil {
+		command.Cancel = func() error {
+			if command.Process == nil {
+				return os.ErrProcessDone
+			}
+			err := syscall.Kill(-command.Process.Pid, syscall.SIGKILL)
+			if err == syscall.ESRCH {
+				return os.ErrProcessDone
+			}
+			return err
+		}
+	}
+	return command
 }
 
 func considerPid(pid int) bool {
@@ -197,9 +199,7 @@ func getSignalsToSend(childProcessName string, noStopExes []string, noKillExes [
 }
 
 func doesChildHaveMarker(_ *psutil.Process, processHandle int) (hasMarker bool, err error) {
-	// the process handle is the pid
-	envCmd := fmt.Sprintf("xargs -0 -n 1 < %s/%d/environ", Procd, processHandle)
-	env, err := exec.Command(getShell(), "-c", envCmd).Output()
+	env, err := readProcFile(processHandle, "environ")
 	if err == nil {
 		hasMarker = containsMarker(string(env))
 	} else {
@@ -208,34 +208,30 @@ func doesChildHaveMarker(_ *psutil.Process, processHandle int) (hasMarker bool, 
 	return hasMarker, err
 }
 
-func _signalProcess(pid int, signalStr string) (err error) {
-	pidStr := strconv.Itoa(pid)
-	out, err := exec.Command("kill", signalStr, pidStr).Output()
+func _signalProcess(pid int, signal syscall.Signal) (err error) {
+	err = syscall.Kill(pid, signal)
 	if Verbose {
-		Log.Infof("ran kill %s %s, waiting for response", signalStr, pidStr)
+		Log.Infof("sent %s to process %d", signal, pid)
 		if err != nil {
 			Log.Error(err)
 		}
-		Log.Infof("%s", out)
 	}
 	return err
 }
 
 func killProcess(processRecord ProcessRecord) error {
-	// -9 means send kill signal
-	return _signalProcess(processRecord.pid, "-9")
+	return _signalProcess(processRecord.pid, syscall.SIGKILL)
 }
 
 func signalProcess(processRecord ProcessRecord, signalNum int) (err error) {
 	switch signalNum {
 	case CTRL_C_SIGNAL:
-		// -SIGINT means send Ctrl+C signal
-		err = _signalProcess(processRecord.pid, "-SIGINT")
+		err = _signalProcess(processRecord.pid, syscall.SIGINT)
 	case CTRL_BREAK_SIGNAL:
 		// nothing to do since Ctrl+Break is Windows only
 		err = nil
 	case KILL_SIGNAL:
-		err = pollKillProcess(processRecord)
+		err = killProcess(processRecord)
 	default:
 		err = errors.New("Unexpected signalNum")
 	}